fix: started chang for mm

+++ Holovibes/sources/gui/windows/panels/image_rendering_panel.cc
@@ -452,12 +452,12 @@ void ImageRenderingPanel::set_divide_convolution(const bool value)
     parent_->notify();
 }
 
-void ImageRenderingPanel::set_z_step(double value)
+void ImageRenderingPanel::set_z_step(int value)
 {
     z_step_ = value;
     ui_->ZSpinBox->setSingleStep(value);
 }
 
-double ImageRenderingPanel::get_z_step() { return z_step_; }
+int ImageRenderingPanel::get_z_step() { return z_step_; }
 
 } // namespace holovibes::gui
